+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 * removed subscribtion kind from Yt::URL
 
 * Change auth: account to as: account
+* Change statistics_set to statistics
+* Removed Yt::URL